Move the fields from PatchListSearchValue to an abstract class.
