+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/patches/PatchListSearchValue.java
@@ -1,64 +1,14 @@
 package network.radicle.jetbrains.radiclejetbrainsplugin.patches;
 
-import com.intellij.collaboration.ui.codereview.list.search.ReviewListSearchValue;
-import org.jetbrains.annotations.Nullable;
+import network.radicle.jetbrains.radiclejetbrainsplugin.toolwindow.AbstractReviewListSearchValue;
 
-public class PatchListSearchValue implements ReviewListSearchValue {
-    public String project;
-    public String searchQuery;
-    public String author;
-    public String state;
-    public String tag;
+public class PatchListSearchValue extends AbstractReviewListSearchValue {
 
     public PatchListSearchValue(PatchListSearchValue searchValue) {
-        this.project = searchValue.project;
-        this.searchQuery = searchValue.searchQuery;
-        this.author = searchValue.author;
-        this.state = searchValue.state;
-        this.tag = searchValue.tag;
-    }
-
-    public String getProject() {
-        return project;
-    }
-    public String getAuthor() {
-        return author;
-    }
-
-    public String getTag() {
-        return tag;
-    }
-    public String getState() {
-        return state;
+        super(searchValue.project, searchValue.searchQuery, searchValue.author, searchValue.state, searchValue.tag);
     }
 
     public PatchListSearchValue() {
-    }
 
-    @Override
-    public int getFilterCount() {
-        int count = 0;
-        if (project != null) {
-            count++;
-        }
-        if (searchQuery != null) {
-            count++;
-        }
-        if (author != null) {
-            count++;
-        }
-        if (state != null) {
-            count++;
-        }
-        if (tag != null) {
-            count++;
-        }
-        return count;
-    }
-
-    @Nullable
-    @Override
-    public String getSearchQuery() {
-        return searchQuery;
     }
 }
